Rename: MainPage.tsx -> index.tsx 네이밍 변경
